Added price update condition
In case the product increases its price after decreasing once then again changes back to the same discounted price we will get no notification in the original code. I added the condition that if the case is as above then the database updates the value of price so as to give a notification when there is a discount again. I couldn't run the code as requests_html is not working on mine so please let me know if it works and is effective.